+++ R/zzz.R
@@ -4,7 +4,7 @@
 .onAttach <- function(lib, pkg) {
   # startup message
   msg <- c(
-    paste0("Thanks for using fHMM version ", utils::packageVersion("fHMM")), "!\n",
+    paste0("Thanks for using {fHMM} version ", utils::packageVersion("fHMM")), "!\n",
     "See https://loelschlaeger.de/fHMM for help.\n",
     "Type 'citation(\"fHMM\")' for citing this R package."
   )